requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==8.0.3
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
gitdb==4.0.7
GitPython==3.1.24
hydra-core==0.11.3
idna==2.9
kiwisolver==1.3.2
lmdb==1.2.1
matplotlib==3.4.3
numpy==1.18.5
omegaconf==1.4.1
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.0
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.24.0
scipy==1.4.1
sentry-sdk==1.4.3
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.46.1
typing-extensions==3.10.0.2
urllib3==1.25.9
wandb==0.12.5
yaspin==2.1.0